Add page details view
Add a button that will fetch details about the current page that is
associated with the access token provided.  The Page ID and Page Name
are sometimes requested when a user files a bug report with Facebook.
There is no easy way to get the Page ID without making an API call, so
added it here for convienence.